+++ src/repositories/accounts/dtos/accounts.dto.ts
@@ -1,5 +1,6 @@
 import { IsString, IsNotEmpty, IsMongoId, IsNumber } from 'class-validator';
 import { PartialType } from '@nestjs/swagger';
+import { CardProvider } from '../accounts.interface';
 
 export class CreateAccountDto {
   @IsString()
@@ -8,11 +9,23 @@ export class CreateAccountDto {
 
   @IsString()
   @IsNotEmpty()
+  readonly alias: string;
+
+  @IsString()
+  @IsNotEmpty()
   readonly accountType: string;
+
+  @IsString()
+  @IsNotEmpty()
+  readonly accountProvider: CardProvider;
 
   @IsNumber()
   @IsNotEmpty()
   readonly amount: number;
+
+  @IsString()
+  @IsNotEmpty()
+  readonly terminationFourDigits: string;
 
   @IsString()
   @IsNotEmpty()